refactor: create utils.basemodel submodule
